Auto reload browser page
I have tried to implement
browser-sync, gulp4 to gulp.
