+++ gulpfile.js
@@ -1,55 +1,164 @@
-const gulp = require('gulp4');
-const uglifyes = require('uglify-es');
-const composer = require('gulp-uglify/composer');
-const uglify = composer(uglifyes, console);
-const babel = require('gulp-babel');
-const less = require('gulp-less');
-const del = require('del');
-const rename = require('gulp-rename');
-const cleanCSS = require('gulp-clean-css');
+// Include the required tools used on tasks
+var gulp = require("gulp"),
+  jshint = require("gulp-jshint"),
+  rename = require("gulp-rename"),
+  uglify = require("gulp-uglify"),
+  saveLicense = require("uglify-save-license"),
+  babel = require("gulp-babel"),
+  postcss = require("gulp-postcss"),
+  cleanCSS = require("gulp-clean-css"),
+  cssbeautify = require("gulp-cssbeautify"),
+  autoprefixer = require("autoprefixer"),
+  del = require("del");
+less = require("gulp-less");
 
-let compileLess = (path, name) => {
-    return gulp.src('src/style.less')
-        .pipe(less({ strictMath: true }))
-        .on('error', swallowError)
-        .pipe(cleanCSS())
-        .pipe(rename(name))
-        .pipe(gulp.dest(path));
-};
+// function swallowError(error) {
+//   console.log(error);
+//   this.emit("end");
+// }
+var Server = require("karma").Server;
+var browserSync = require("browser-sync").create();
 
-let compileJs = (path, name) => gulp.src('src/countDown.js')
-    .pipe(babel({
-        presets: ['@babel/env'],
-        plugins: ['@babel/plugin-transform-modules-umd']
-    }))
-    .pipe(uglify({
+// Specify the Source files
+var SRC_JS = "src/js/*.js";
+var SRC_CSS = "src/css/*.css";
+var SRC_LESS = "src/less/*.less";
+
+// Specify the Destination folders
+var DEST_JS = "dist/js";
+var DEST_CSS = "dist/css";
+var DEST_LESS = "dist/css";
+
+// Example pages
+var EXAMPLE_HTML = "examples/*.html";
+
+
+
+function compile_js(cb) {
+  gulp
+    .src(SRC_JS)
+    .pipe(
+      babel({
+        presets: ["@babel/env"],
+        plugins: ["@babel/plugin-transform-modules-umd"],
+      })
+    )
+    .pipe(
+      uglify({
         mangle: true,
-        compress: true
-    }))
-    .pipe(rename(name))
-    .pipe(gulp.dest(path));
+        compress: true,
+      })
+    )
+    .pipe(rename("countDown.min.js"))
+    .pipe(gulp.dest(DEST_JS));
 
-function swallowError(error) {
-    console.log(error);
-    this.emit('end');
+  cb();
 }
 
-gulp.task('js', () => compileJs('build/', 'countDown.js'));
-gulp.task('js:watch', () => gulp.watch('src/countDown.js'), gulp.series('js'));
+// Compline less
+function compile_less(cb) {
+  gulp
+    .src(SRC_LESS)
+    .pipe(less({ strictMath: true }))
+    // .on("error", swallowError)
+    .pipe(cleanCSS())
+    .pipe(rename("style.css"))
+    .pipe(gulp.dest(DEST_LESS));
+  cb();
+}
 
-gulp.task('less', () => compileLess('build/', 'style.css'));
-gulp.task('less:watch', () => gulp.watch('./src/style.less', gulp.series('less')));
+// BUILD CSS
+function build_css(cb) {
+  gulp
+    .src(SRC_CSS)
+    .pipe(postcss([autoprefixer()]))
+    .pipe(cssbeautify({ autosemicolon: true }))
+    .pipe(gulp.dest(DEST_CSS))
+    .pipe(cleanCSS({ compatibility: "ie8" }))
+    .pipe(rename({ suffix: ".min" }))
+    .pipe(gulp.dest(DEST_CSS));
+  cb();
+}
 
-gulp.task('example:less', () => compileLess('example/', 'style.css'));
-gulp.task('example:compile_js', () => compileJs('example/', 'countDown.min.js'));
+// LINT
+function lint_js(cb) {
+  gulp
+    .src(SRC_JS)
+    .pipe(jshint({ esversion: 8 }))
+    .pipe(jshint.reporter("default"));
+  cb();
+}
 
-gulp.task('dist:less', () => compileLess('dist/', 'countDown.min.css'));
-gulp.task('dist:compile_js', () => compileJs('dist/', 'countDown.min.js'));
+// CLEAN
+function clean_js(cb) {
+  del.sync([DEST_JS]);
+  cb();
+}
 
-gulp.task('develop', gulp.parallel(gulp.series('less', 'less:watch'), gulp.series('js', 'js:watch')));
+function clean_css(cb) {
+  del.sync([DEST_CSS]);
+  cb();
+}
 
-gulp.task('update:example', gulp.parallel('example:less', 'example:compile_js'));
-gulp.task('update:dist', gulp.parallel('dist:less', 'dist:compile_js'));
-gulp.task('update:all', gulp.parallel('update:dist', 'update:example'));
+function clean_less(cb) {
+  del.sync([DEST_LESS]);
+  cb();
+}
 
-gulp.task('clean', () => del(['build/**/*',]));
+// WATCH
+function watch(cb) {
+  // gulp.watch(SRC_JS, build_js);
+  gulp.watch(SRC_CSS, build_css);
+  gulp.watch(SRC_LESS, compile_less);
+  gulp.watch(SRC_JS, compile_js);
+
+  cb();
+}
+
+// SERVE
+function serve(cb) {
+  // Serve files from the root of this project
+  browserSync.init({
+    server: {
+      baseDir: ["examples", "dist"],
+      index: "index.html",
+    },
+  });
+
+  // gulp.watch(SRC_JS, build_js);
+  gulp.watch(SRC_CSS, build_css);
+  gulp.watch(SRC_LESS, compile_less);
+  gulp.watch(SRC_JS, compile_js);
+
+  gulp
+    .watch([DEST_JS, DEST_CSS, SRC_LESS, EXAMPLE_HTML])
+    .on("change", browserSync.reload);
+
+  cb();
+}
+
+// TEST
+function test(cb) {
+  new Server(
+    {
+      configFile: __dirname + "/karma.conf.js",
+      singleRun: true,
+    },
+    done
+  ).start();
+
+  cb();
+}
+
+// EXPORT methods
+exports.clean = gulp.parallel(clean_js, clean_css, clean_less);
+exports.build = gulp.parallel(
+  gulp.series(clean_js, lint_js, compile_js),
+  gulp.series(clean_css, build_css),
+  gulp.series(clean_less, compile_less)
+);
+exports.lint = lint_js;
+exports.watch = watch;
+exports.test = test;
+exports.serve = serve;
+exports.default = serve;
